nuci-tls: The CA expiration is also 10 years
It makes little sense to have the clients for 10 years and the CA for
single month only. Set it to 10 years as well.

It doesn't seem to be possible to set it in the config file when it is
just self-signed cert (at least no way was found), so it is passed on
the command line.